# apps/api/app/main.py
from __future__ import annotations

import atexit
import logging
import os
from contextlib import asynccontextmanager
from pathlib import Path
from typing import AsyncIterator

from fastapi import FastAPI, HTTPException
from fastapi.exceptions import RequestValidationError
from fastapi.middleware.cors import CORSMiddleware
from fastapi.responses import RedirectResponse
from fastapi.staticfiles import StaticFiles

from .api.dependencies import build_runtime
from .api.error_handlers import agentic_saffron_error_handler, http_exception_handler, validation_exception_handler
from .api.routes.agent import router as agent_router
from .api.routes.audit import router as audit_router
from .api.routes.auth import router as auth_router
from .api.routes.briefings import router as briefings_router
from .api.routes.chat import router as chat_router
from .api.routes.data import router as data_router
from .api.routes.documents import router as documents_router
from .api.routes.health import router as health_router
from .api.routes.ingestion import router as ingestion_router
from .api.routes.intelligence import router as intelligence_router
from .api.routes.intelligence_map import router as intelligence_map_router
from .api.routes.platform import router as platform_router
from .api.routes.research import router as research_router
from .api.routes.sources import router as sources_router
from .api.routes.voice import router as voice_router
from .config.settings import AppSettings
from .domain.errors import AgenticSaffronError
from .middleware.rate_limit import RateLimitMiddleware
from .middleware.request_id import RequestIdMiddleware
from .middleware.request_size import RequestSizeLimitMiddleware
from .middleware.security_headers import SecurityHeadersMiddleware
from .middleware.timeout import RequestTimeoutMiddleware

settings = AppSettings.from_env()


def _configure_logging(level_name: str | None) -> None:
    """Show the app's own log lines at SAFFRON_LOG_LEVEL (for example INFO); unset keeps Python's warnings-only default."""

    level = logging.getLevelName((level_name or "").strip().upper()) if level_name else None
    if not isinstance(level, int):
        return
    handler = logging.StreamHandler()
    handler.setFormatter(logging.Formatter("%(levelname)s %(name)s: %(message)s"))
    for name in ("saffron", "app"):
        app_logger = logging.getLogger(name)
        if not app_logger.handlers:
            app_logger.addHandler(handler)
        app_logger.setLevel(level)
        # Handled here; the root logger's last-resort handler would print it again.
        app_logger.propagate = False


_configure_logging(os.getenv("SAFFRON_LOG_LEVEL"))

@asynccontextmanager
async def lifespan(app: FastAPI) -> AsyncIterator[None]:
    yield
    app.state.runtime.close()

app = FastAPI(title=settings.app_name, description="AI-powered institutional intelligence platform: ingestion, canonical data, policy-bound agents, and internet intelligence", version=settings.version, lifespan=lifespan)
app.state.runtime = build_runtime(settings, start_workers=True)  # the API process runs the thread queue, if configured
atexit.register(app.state.runtime.close)
app.add_exception_handler(AgenticSaffronError, agentic_saffron_error_handler)  # type: ignore[arg-type]
app.add_exception_handler(HTTPException, http_exception_handler)  # type: ignore[arg-type]
app.add_exception_handler(RequestValidationError, validation_exception_handler)  # type: ignore[arg-type]
app.add_middleware(RequestIdMiddleware)
app.add_middleware(SecurityHeadersMiddleware, production=settings.environment == "production", auth_origins=settings.oidc_browser_origins)
app.add_middleware(RequestTimeoutMiddleware, timeout_seconds=settings.request_timeout_seconds)
app.add_middleware(RateLimitMiddleware, max_requests=settings.rate_limit_requests, window_seconds=settings.rate_limit_window_seconds)
app.add_middleware(RequestSizeLimitMiddleware, max_bytes=settings.max_request_bytes, upload_max_bytes=settings.max_upload_bytes + 65_536, upload_routes=(("POST", "/v1/ingestion/uploads"), ("POST", "/v1/documents")))
if settings.allowed_origins:
    app.add_middleware(CORSMiddleware, allow_origins=list(settings.allowed_origins), allow_credentials=True, allow_methods=["GET", "POST", "DELETE", "OPTIONS"], allow_headers=["*"])
app.include_router(health_router)
app.include_router(auth_router)
app.include_router(chat_router)
app.include_router(briefings_router)
app.include_router(voice_router)
app.include_router(sources_router)
app.include_router(research_router)
app.include_router(audit_router)
app.include_router(ingestion_router)
app.include_router(data_router)
app.include_router(agent_router)
app.include_router(documents_router)
app.include_router(intelligence_router)
app.include_router(intelligence_map_router)
app.include_router(platform_router)

# Two separate browser apps: the client assistant at `/` and the developer
# platform console at `/console/`. They share only the sign-in client and base
# styles under `/shared/`; neither links to the other. The console is mounted
# only when enabled, so a client-facing deployment does not serve it at all.
web_root = Path(__file__).resolve().parents[2] / "web"
if web_root.exists():
    app.mount("/shared", StaticFiles(directory=web_root / "shared"), name="web-shared")
    if settings.web_console_enabled:

        @app.get("/console", include_in_schema=False)
        @app.get("/platform.html", include_in_schema=False)  # the console's previous address
        async def console_entry() -> RedirectResponse:
            # The `/` mount below would otherwise answer these paths with a 404.
            return RedirectResponse("/console/")

        app.mount("/console", StaticFiles(directory=web_root / "console", html=True), name="web-console")
    app.mount("/", StaticFiles(directory=web_root / "assistant", html=True), name="web-assistant")

__all__ = ["app"]