# openapi.yaml
openapi: 3.1.0
info:
  title: agentic-saffron-api
  description: 'AI-powered institutional intelligence platform: ingestion, canonical
    data, policy-bound agents, and internet intelligence'
  version: 0.1.0
paths:
...
  /v1/ingestion/uploads:
    post:
      tags:
      - ingestion
      summary: Upload an existing institutional file (Excel, CSV, PDF, Word, scan,
        JSON)
      operationId: upload_file_v1_ingestion_uploads_post
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_upload_file_v1_ingestion_uploads_post'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Upload File V1 Ingestion Uploads Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /v1/documents:
    post:
      tags:
      - documents
      summary: Upload a policy, circular, or other document for retrieval
      operationId: upload_document_v1_documents_post
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_upload_document_v1_documents_post'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Upload Document V1 Documents Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - documents
      summary: List indexed documents visible to the caller
      operationId: list_documents_v1_documents_get
      parameters:
      - name: institution_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Institution Id
      - name: category
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Category
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 50
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response List Documents V1 Documents Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    Body_upload_document_v1_documents_post:
      properties:
        file:
          type: string
          contentMediaType: application/octet-stream
          title: File
        title:
          anyOf:
          - type: string
          - type: 'null'
          title: Title
        classification:
          type: string
          title: Classification
          default: internal
        category:
          type: string
          title: Category
          default: general
        institution_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Institution Id
      type: object
      required:
      - file
      title: Body_upload_document_v1_documents_post
    Body_upload_file_v1_ingestion_uploads_post:
      properties:
        file:
          type: string
          contentMediaType: application/octet-stream
          title: File
        entity:
          anyOf:
          - type: string
          - type: 'null'
          title: Entity
        sheet:
          anyOf:
          - type: string
          - type: 'null'
          title: Sheet
        institution_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Institution Id
        auto_commit:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Auto Commit
        process:
          type: boolean
          title: Process
          default: true
      type: object
      required:
      - file
      title: Body_upload_file_v1_ingestion_uploads_post
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...